ALLOCATE support in Parser and Sema
Add types for allocation statement and array expression. Do basic verification
for the array alocation semantics in ALLOCATE. Minor additition to the test.

Create a more general type for allocation expression (instead of array
allocation), but leave other cases unimplemented for now. Type of the
allocation is the type of the target expression, which may not be correct in
all cases.

Add Sema support for ALLOCATE statements.

For #22